+++ projects/ngx-hateoas-client/src/lib/model/resource/resource.ts
@@ -51,14 +51,18 @@ export class Resource extends BaseResource {
   }
 
   /**
-   * Adds the passed entity to the resource collection behind the relation name.
+   * Adding passed entities to the resource collection behind the relation name.
+   * Used POST method with 'Content-Type': 'text/uri-list'.
+   *
+   * This method DOES NOT REPLACE existing resources in the collection instead it adds new ones.
+   * To replace collection resource with passed entities use {@link bindRelation} method.
    *
    * @param relationName used to get the specific resource relation link to the resource collection
    * @param entities one or more entities that should be added to the resource collection
    * @throws error when required params are not valid or link not found by relation name
    */
-  public addRelation<T extends Resource>(relationName: string, entities: Array<T>): Observable<HttpResponse<any>> {
-    StageLogger.resourceBeginLog(this, 'ADD_RELATION', {relationName, resourceLinks: this._links, entities});
+  public addCollectionRelation<T extends Resource>(relationName: string, entities: Array<T>): Observable<HttpResponse<any>> {
+    StageLogger.resourceBeginLog(this, 'ADD_COLLECTION_RELATION', {relationName, resourceLinks: this._links, entities});
     ValidationUtils.validateInputParams({relationName, entities});
 
     const relationLink = this.getRelationLink(relationName);
@@ -67,68 +71,82 @@ export class Resource extends BaseResource {
       .map(entity => {
         return ResourceUtils.initResource(entity).getSelfLinkHref();
       })
-      .join('\n\r');
+      .join('\n');
 
     return getResourceHttpService().post(UrlUtils.generateLinkUrl(relationLink), body, {
       observe: 'response',
       headers: new HttpHeaders({'Content-Type': 'text/uri-list'})
     }).pipe(
       tap(() => {
-        StageLogger.resourceEndLog(this, 'ADD_RELATION', {result: `relation ${ relationName } was added successful`});
+        StageLogger.resourceEndLog(this, 'ADD_COLLECTION_RELATION',
+          {result: `collection relation ${ relationName } was updated successfully`});
       })
     );
   }
 
   /**
-   * Update exist resource relation value.
+   * Bounding the passed entity or collection of entities to this resource by the relation name.
+   * Used PUT method with 'Content-Type': 'text/uri-list'.
    *
-   * @param relationName with which should be updated
-   * @param entity new value of entity
-   * @throws error when required params are not valid or link not found by relation name
-   */
-  public updateRelation<T extends Resource>(relationName: string, entity: T): Observable<HttpResponse<any>> {
-    StageLogger.resourceBeginLog(this, 'UPDATE_RELATION', {relationName, resourceLinks: this._links, entity});
-    ValidationUtils.validateInputParams({relationName, entity});
-
-    const relationLink = this.getRelationLink(relationName);
-    const resource = ResourceUtils.initResource(entity) as Resource;
-
-    return getResourceHttpService().patch(UrlUtils.generateLinkUrl(relationLink), resource.getSelfLinkHref(), {
-      observe: 'response',
-      headers: new HttpHeaders({'Content-Type': 'text/uri-list'})
-    }).pipe(
-      tap(() => {
-        StageLogger.resourceEndLog(this, 'UPDATE_RELATION', {result: `relation ${ relationName } was bind successful`});
-      })
-    );
-  }
-
-  /**
-   * Bind the passed entity to this resource by the relation name.
+   * This method also REPLACED existing resources in the collection by passed entities.
+   * To add entities to collection resource use {@link addCollectionRelation} method.
    *
    * @param relationName with which will be associated passed entity to this resource
-   * @param entity that should be bind to this resource
+   * @param entities one or more entities that should be bind to this resource
    * @throws error when required params are not valid or link not found by relation name
    */
-  public bindRelation<T extends Resource>(relationName: string, entity: T): Observable<HttpResponse<any>> {
-    StageLogger.resourceBeginLog(this, 'BIND_RELATION', {relationName, resourceLinks: this._links, entity});
-    ValidationUtils.validateInputParams({relationName, entity});
+  public bindRelation<T extends Resource>(relationName: string, entities: Array<T>): Observable<HttpResponse<any>> {
+    StageLogger.resourceBeginLog(this, 'BIND_RELATION', {relationName, resourceLinks: this._links, entities});
+    ValidationUtils.validateInputParams({relationName, entities});
 
     const relationLink = this.getRelationLink(relationName);
-    const resource = ResourceUtils.initResource(entity) as Resource;
+    const body = entities
+      .map(entity => {
+        return ResourceUtils.initResource(entity).getSelfLinkHref();
+      })
+      .join('\n');
 
-    return getResourceHttpService().put(UrlUtils.generateLinkUrl(relationLink), resource.getSelfLinkHref(), {
+    return getResourceHttpService().put(UrlUtils.generateLinkUrl(relationLink), body, {
       observe: 'response',
       headers: new HttpHeaders({'Content-Type': 'text/uri-list'})
     }).pipe(
       tap(() => {
-        StageLogger.resourceEndLog(this, 'BIND_RELATION', {result: `relation ${ relationName } was bind successful`});
+        StageLogger.resourceEndLog(this, 'BIND_RELATION', {result: `relation ${ relationName } was bound successfully`});
+      })
+    );
+  }
+
+  /**
+   * Unbinding single resource relation behind resource name.
+   * Used DELETE method to relation resource link URL.
+   *
+   * This method DOES NOT WORK WITH COLLECTION RESOURCE relations.
+   * To clear collection resource relation use {@link clearCollectionRelation} method.
+   * To delete one resource from resource collection use {@link deleteRelation} method.
+   *
+   * @param relationName resource relation name to unbind
+   */
+  public unbindRelation<T extends Resource>(relationName: string): Observable<HttpResponse<any>> {
+    StageLogger.resourceBeginLog(this, 'UNBIND_RELATION', {relationName, resourceLinks: this._links});
+    ValidationUtils.validateInputParams({relationName});
+
+    const relationLink = this.getRelationLink(relationName);
+
+    return getResourceHttpService().delete(UrlUtils.generateLinkUrl(relationLink), {
+      observe: 'response',
+    }).pipe(
+      tap(() => {
+        StageLogger.resourceEndLog(this, 'UNBIND_RELATION', {result: `relation ${ relationName } was unbound successfully`});
       })
     );
   }
 
   /**
    * Unbind all resources from collection by the relation name.
+   * Used PUT method with 'Content-Type': 'text/uri-list' and EMPTY body to clear relations.
+   *
+   * To delete one resource from collection use {@link deleteRelation} method.
+   * To delete single resource relations use {@link unbindRelation} or {@link deleteRelation} methods.
    *
    * @param relationName used to get relation link to unbind
    * @throws error when required params are not valid or link not found by relation name
@@ -144,19 +162,20 @@ export class Resource extends BaseResource {
       headers: new HttpHeaders({'Content-Type': 'text/uri-list'})
     }).pipe(
       tap(() => {
-        StageLogger.resourceEndLog(this, 'CLEAR_COLLECTION_RELATION', {result: `relation ${ relationName } was cleared successful`});
+        StageLogger.resourceEndLog(this, 'CLEAR_COLLECTION_RELATION', {result: `relation ${ relationName } was cleared successfully`});
       })
     );
   }
 
   /**
-   * Unbind the single resource by the relation name.
+   * Deleting resource relation.
+   * For collection, means that only passed entity will be unbound from the collection.
+   * For single resource, deleting relation the same as @{link unbindRelation} method.
    *
-   * If behind relation name is link to collection of the resources then it means that
-   * only passed entity will be unbind from collection of the resource.
+   * To delete all resource relations from collection use {@link clearCollectionRelation} method.
    *
    * @param relationName used to get relation link to unbind
-   * @param entity that should be unbind from this resource
+   * @param entity that should be unbind from this relation
    * @throws error when required params are not valid or link not found by relation name
    */
   public deleteRelation<T extends Resource>(relationName: string, entity: T): Observable<HttpResponse<any>> {
@@ -185,7 +204,7 @@ export class Resource extends BaseResource {
       observe: 'response'
     }).pipe(
       tap(() => {
-        StageLogger.resourceEndLog(this, 'DELETE_RELATION', {result: `relation ${ relationName } was deleted successful`});
+        StageLogger.resourceEndLog(this, 'DELETE_RELATION', {result: `relation ${ relationName } was deleted successfully`});
       })
     );
   }