+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.1.0 (2021-02-21)
+#### Updated Resource methods 
+
+Deleted `Resource` methods:
+- `updateResource`
+
+Changed `Resource` methods:
+- `addRelation` was renamed to `addCollectionRelation`
+- `bindRelation` changed signature method now it accepts an array of entities instead single entity.
+
+Added `Resource` methods:
+- `unbindResource` a new method that used with single resource relations to delete bound relations.
+
+README.md was updated.
+
 ## 1.0.1 (2021-02-09)
-#### Update angular
+#### Updated angular
 Updated Angular to 11 version.